working  on delete_blog

+++ src/main/service/blog_service.py
@@ -1,3 +1,5 @@
+from bson import ObjectId
+
 from src.main.data.model.blog import Blog
 from src.main.data.repo.blog_storage import BlogStorage
 
@@ -13,11 +15,18 @@ class BlogService:
             return {"error":"Field Required"},401
         if blog.content is None:
             return {"error":"Field Required"},401
-        self.storage.save_blog(blog)
-        return {"message":"Blog added successfully"},200
+        result = self.storage.save_blog(blog)
+        blog_id = str(result.inserted_id)
+        return{
+            "blog_id": blog_id,
+            "message": "Blog added successfully",
+            "status": 200
+        }
 
 
     def delete_blog(self, blog_id):
+        if not ObjectId.is_valid(blog_id):
+            return {"error": "Invalid blog ID"}, 400
         if self.storage.exist_by_blog_id(blog_id):
             self.storage.delete_blog(blog_id)
             return {"message":"Blog deleted successfully"},200
@@ -25,5 +34,8 @@ class BlogService:
 
     def find_blog(self, blog_id):
         if self.storage.exist_by_blog_id(blog_id):
-            return self.storage.find_blog_by_id(blog_id)
-        return {"error":"Blog not found"},404
+            blog = self.storage.find_blog_by_id(str(blog_id))
+            return blog, 200
+        return {"error": "Blog not found"}, 404
+
+
